+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Version 2.0.2 (unreleased)
 
 - Fixed static analysis of filters in ternary expressions. See [#180](https://github.com/jg-rp/liquid/issues/180).
+- Fixed static analysis of macro blocks. Previously `args` and `kwargs` were considered "global". See [#181](https://github.com/jg-rp/liquid/issues/181).
 
 ## Version 2.0.1